+++ src/reader/__init__.py
@@ -79,6 +79,7 @@ from .exceptions import (
     ResourceNotFoundError,
     PluginError,
     InvalidPluginError,
+    ReaderWarning,
 )